riscv64: Add SyncupEnv & TooManyIR jump kinds
SyncupEnv is added for syncup the environment so that the following
instructions can get the update from the previous instructions, e.g.
vl set by vsetvl.

TooManyIR is added for the cases one guest instruction is translated to
many IRs, this stops TB including more guest instructions and it is used
to avoid too many IRs in one TB.

Signed-off-by: Fei Wu <fei2.wu@intel.com>